backend/requirements.txt
fastapi
uvicorn
pymupdf
python-multipart
sentence-transformers
faiss-cpu
numpy
requests
torch --index-url https://download.pytorch.org/whl/cpu